services/extract/requirements.txt
google-cloud-storage==3.3.1
numpy==2.2.5
pandas==2.2.3
pyarrow==21.0.0
pydantic==2.11.7
requests==2.32.4
yfinance==0.2.65